Change to linear indices for the boundary lists
Both the real and ghost node indices in the boundary indices lists are
now stored as linear indices instead of cartesian indices. This was done
mostly for the very minor performance gains (I was hoping it would solve
something else, which it didn't, but I thought I might as well keep it
for the small speed up).

The cartesian index of the real node is still included, and is what is
still what is passed into the index field of the
boundaryHeatTransferRate function, so no user facing things should
change.