rename innerLoop! to ghostCalc!, update the docs and rename some of its variables
The function no longer exists just as a barrier function, so it has been renamed to
reflect that. The note about barrier functions has been moved from the interface to
an internal comment, and the comments in padWithGhost have also been cleaned up.
The indices in ghostCalc! were also renamed to help clear up what they actually are.